revert lambda changes
