+++ .vscode/launch.json
@@ -9,7 +9,7 @@
             "type": "go",
             "request": "launch",
             "mode": "debug",
-            "program": "${workspaceRoot}/cmd/http/main.go",
+            "program": "${workspaceRoot}/main.go",
             "envFile": "${workspaceFolder}/.env"
         }
     ]
