+++ Dockerfile
@@ -7,16 +7,16 @@ ARG BASE_PATH
 COPY . ${BASE_PATH}
 WORKDIR ${BASE_PATH}
 
-RUN GO111MODULE=on CGO_ENABLED=0 go build -o randflix-lambda ./cmd/lambda
+RUN GO111MODULE=on CGO_ENABLED=0 go build -o randflix-api .
 
-FROM public.ecr.aws/lambda/go:1 AS final
+FROM scratch AS final
 
 ARG BASE_PATH
 
 # Copy ca certs from build container
 COPY --from=build /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/
 # Copy application
-COPY --from=build ${BASE_PATH}/randflix-lambda /randflix-lambda
+COPY --from=build ${BASE_PATH}/randflix-api /randflix-api
 
 EXPOSE 8080
-CMD [ "/randflix-lambda" ]
+ENTRYPOINT [ "/randflix-api" ]
