+++ go.mod
@@ -3,7 +3,6 @@ module github.com/microhod/randflix-api
 go 1.15
 
 require (
-	github.com/akrylysov/algnhsa v0.12.1
 	github.com/gobeam/mongo-go-pagination v0.0.2
 	github.com/gorilla/mux v1.8.0
 	github.com/kelseyhightower/envconfig v1.4.0
